Fix headder to publish banner
